processors/sentiment/requirements.txt
functions-framework==3.5.0
anthropic==0.39.0
psycopg2-binary==2.9.9
